% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choladd.R
\name{choladd}
\alias{choladd}
\title{Cholesky Row/Column Addition}
\usage{
choladd(CF, z, k, lower = TRUE)
}
\arguments{
\item{CF}{Cholesky factor, either lower-triangular (in which case set
`lower = TRUE`) or upper-triangular (in which case set `upper = FALSE`)}

\item{z}{The "missing" row/column of `M`, i.e. any `p + 1`-long vector which,
upon being inserted to be the `k`-th row/column in expanded matrix `M`,
retains positive definiteness `M > 0`}

\item{k}{The row/column index which `z`'s insertion will correspond to,
from `1` to `p + 1` inclusive}

\item{lower}{Boolean, dictating whether `CF` is the lower Cholesky factor
(`TRUE`, default) or upper Cholesky factor (`FASLE`). Also dictates whether
the returned Choleskyfactor is lower- or upper-triangular.}
}
\value{
Cholesky factor of matrix `M`, where `M` is constructed from
`M[-k, -k]` and `z` inserted at the `k`-th row/column index.
}
\description{
This function calculates the lower/upper Cholesky factor of positive definite
matrix `M`, provided the Cholesky factor of `M[-k, -k]` is known. That is,
this function inverts `choldrop`, re-inserting the `k`^th row/column from
`M`.
}
\examples{
# Generate positive definite matrix
p = 20
M = diag(p) + tcrossprod(matrix(rnorm(p * 2), p, 2))

# Take out a row/column of M
k = 6
Mk = M[-k, -k]
x = M[ , k]

# Calculate Cholesky factor of Mk
Uk = chol(Mk)

# Calculate Cholesky factor adding back in the k-th row/column
L = choladd(t(Uk), k, lower = TRUE)

# Check against Cholesky factor of Mk
max(abs(L - t(chol(M))))
}
