% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choladdL.R
\name{choladdL}
\alias{choladdL}
\title{Lower Cholesky Row/Column Addition}
\usage{
choladdL(L, z, k)
}
\arguments{
\item{L}{Lower Cholesky factor of dimension `p` \eqn{\times} `p`, such that
`LL' = M[-k, -k]`}

\item{z}{The "missing" row/column of `M`, i.e. any `p + 1`-long vector which,
upon being inserted to be the `k`-th row/column in expanded matrix `M`,
retains positive definiteness `M > 0`}

\item{k}{The row/column index which `z`'s insertion will correspond to,
from `1` to `p + 1` inclusive}
}
\value{
Lower Cholesky factor of matrix `M`, where `M` is constructed from
`M[-k, -k]` and `z` inserted at the `k`-th row/column index.
}
\description{
This function calculates the lower Cholesky factor of positive definite
matrix `M`, provided lower Cholesky factor of `M[-k, -k]` is known. That is,
this function inverts `choldropL`, re-inserting the `k`^th row/column from
`M`.
}
