% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choldownL.R
\name{choldownL}
\alias{choldownL}
\title{Lower Cholesky Rank-1 Downdate}
\usage{
choldownL(L, x)
}
\arguments{
\item{L}{Lower Cholesky factor of dimension `p` \eqn{\times} `p`}

\item{x}{A `p`-long vector such that `LL' - xx' > 0`}
}
\value{
Lower Cholesky factor of `LL' - xx'`
}
\description{
This function calculates the rank-1 downdate of lower Cholesky factor `L`,
i.e. the lower Cholesky factor of matrix `LL' - xx'` (provided that matrix is
also positive definite). Here, `L` is a square `p` \eqn{\times} `p`
lower-triangular matrix with positive diagonal entries, and `x` is any vector
of length `p` such that `LL' - xx' > 0`. This function applies hyperbolic
rotations to triangularize the joint matrix `(L x)`. Due to cache latency,
this function will generally be faster than `choldownU`.
}
