% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choldownU.R
\name{choldownU}
\alias{choldownU}
\title{Upper Cholesky Rank-1 Downdate}
\usage{
choldownU(U, x)
}
\arguments{
\item{U}{Upper Cholesky factor of dimension `p` \eqn{\times} `p`}

\item{x}{A `p`-long vector such that `U'U - xx' > 0`}
}
\value{
Upper Cholesky factor of `U'U - xx'`
}
\description{
This function calculates the rank-1 downdate of upper Cholesky factor `U`,
i.e. the upper Cholesky factor of matrix `U'U - xx'` (provided that matrix is
also positive definite). Here, `U` is a square `p` \eqn{\times} `p`
upper-triangular matrix with positive diagonal entries, and `x` is any vector
of length `p` such that `U'U - xx' > 0`. This function applies hyperbolic
rotations to triangularize the joint matrix `(U x)`. Due to cache latency,
this function will generally be slower than `choldownL`.
}
