% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choldropL.R
\name{choldropL}
\alias{choldropL}
\title{Lower Cholesky Row/Column Drop}
\usage{
choldropL(L, k)
}
\arguments{
\item{L}{Lower Cholesky factor of dimension `p` \eqn{\times} `p`, such that
`LL' = M`}

\item{k}{An index between `1` and `p` inclusive, corresponding to the
row/column to be dropped}
}
\value{
Lower Cholesky factor of `M[-k, -k]`, i.e. of `M` with the
\eqn{k^{th}} row/column removed
}
\description{
This function calculates the lower Cholesky factor of matrix `M[-k, -k]`,
i.e. of positive definite `M` with the \eqn{k^{th}} row/column removed,
provided lower Cholesky factor `L` (`M = LL'`) is known. Due to cache
latency, this function will generally be faster than `choldropU`.
}
