% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choldropU.R
\name{choldropU}
\alias{choldropU}
\title{Uper Cholesky Row/Column Drop}
\usage{
choldropU(U, k)
}
\arguments{
\item{U}{Upper Cholesky factor of dimension `p` \eqn{\times} `p`, such that
`U'U = M`}

\item{k}{An index between `1` and `p` inclusive, corresponding to the
row/column to be dropped}
}
\value{
Upper Cholesky factor of `M[-k, -k]`, i.e. of `M` with the
\eqn{k^{th}} row/column removed
}
\description{
This function calculates the upper Cholesky factor of matrix `M[-k, -k]`,
i.e. of positive definite `M` with the \eqn{k^{th}} row/column removed,
provided upper Cholesky factor `U` (`M = U'U`) is known. Due to cache
latency, this function will generally be slower than `choldropL`.
}
