% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholRcpp.R
\name{cholRcpp}
\alias{cholRcpp}
\title{Cholesky Decomposition with Rcpp}
\usage{
cholRcpp(M, lower = TRUE)
}
\arguments{
\item{M}{Any square positive definite matrix}

\item{lower}{Boolean, dictating whether the lower Cholesky factor (`TRUE`,
default) or upper Cholesky factor (`FASLE`) should be returned}
}
\value{
The lower (or upper) Cholesky factor of `M`
}
\description{
This function calculates the Cholesky decomposition of matrix `M` with the
Rcpp library implementation, which permits calculating either the lower or
upper Cholesky factors. Calculating the lower Cholesky factor tends to be
twice as fast, so it is the default.
}
