% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholup.R
\name{cholup}
\alias{cholup}
\title{Cholesky Rank-1 Update}
\usage{
cholup(CF, x, lower = TRUE)
}
\arguments{
\item{CF}{Cholesky factor of dimension `p` \eqn{\times} `p`}

\item{x}{Any `p`-long vector}
}
\value{
Cholesky factor of (e.g.) `(CF)(CF') + xx'`, where `(CF)(CF')` is
the positive definite matrix associated with `CF`.
}
\description{
This function calculates the rank-1 update of Cholesky factor `CF`, e.g.
the Cholesky factor of matrix `(CF)(CF') + xx'`. Here, `CF` is a square `p`
\eqn{\times} `p` Cholesky factor, and `x` is any arbitrary vector of length
`p`. The user may specify whether `CF` is lower-triangular (in which case
`lower = TRUE`) or upper-triangular (in which case `lower = FALSE`). This
function applies Givens rotations to triangularize the joint matrix `(CF x)`;
due to cache latency, it will generally be faster when `CF` is the lower
Cholesky factor.
}
\examples{
# Generate positive definite matrix
p = 20
M = diag(p) + tcrossprod(matrix(rnorm(p * 2), p, 2))

# Generate vector for rank-1 update
x = rnorm(p)

# Calculate Cholesky factor of M
U = chol(M)

# Calculate rank-1 update
Ux = cholup(U, x, lower = FALSE)

# Check against Cholesky factor of U'U + xx'
Mx = M + tcrossprod(x)
max(abs(Ux - chol(Mx)))
}
