% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholupK.R
\name{cholupK}
\alias{cholupK}
\title{Cholesky Rank-K Update}
\usage{
cholupK(CF, X, lower = TRUE)
}
\arguments{
\item{CF}{Cholesky factor of dimension `p` \eqn{\times} `p`}

\item{X}{Any `p` \eqn{\times} `K` matrix}

\item{lower}{Boolean, dictating whether `CF` is the lower Cholesky factor
(`TRUE`, default) or upper Cholesky factor (`FASLE`). Also dictates whether
the returned Choleskyfactor is lower- or upper-triangular.}
}
\value{
Cholesky factor of (e.g.) `(CF)(CF') + xx'`, where `(CF)(CF')` is
the positive definite matrix associated with `CF`.
}
\description{
This function calculates the rank-1 update of Cholesky factor `CF`, e.g.
the Cholesky factor of matrix `(CF)(CF') + xx'`. Here, `CF` is a square `p`
\eqn{\times} `p` Cholesky factor, and `x` is any arbitrary vector of length
`p`. The user may specify whether `CF` is lower-triangular (in which case
`lower = TRUE`) or upper-triangular (in which case `lower = FALSE`). This
function applies Givens rotations to triangularize the joint matrix `(CF x)`;
due to cache latency, it will generally be faster when `CF` is the lower
Cholesky factor.
}
\examples{
# Generate positive definite matrix
p = 20
M = diag(p) + tcrossprod(matrix(rnorm(p * 2), p, 2))

# Generate matrix for rank-K update, i.e. M + XX' = M + sum_k {X[ , k] * X[ , k]'}
K = 3
X = matrix(rnorm(p * K), p, K)

# Calculate Cholesky factor of M
U = chol(M)

# Calculate rank-1 update
UX = cholupK(U, X, lower = FALSE)

# Check against Cholesky factor of U'U + XX'
MX = M + tcrossprod(X)
max(abs(UX - chol(MX)))
}
