% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholupL.R
\name{cholupL}
\alias{cholupL}
\title{Cholesky Rank-1 Update Using Lower Cholesky Factor L}
\usage{
cholupL(L, x)
}
\arguments{
\item{L}{Lower-Cholesky factor of dimension `p` \eqn{\times} `p`}

\item{x}{Any `p`-long vector}
}
\value{
Matrix `M = L + xx'`
}
\description{
Cholesky Rank-1 Update Using Lower Cholesky Factor L
}
