% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholupU.R
\name{cholupU}
\alias{cholupU}
\title{Upper Cholesky Rank-1 Update}
\usage{
cholupU(U, x)
}
\arguments{
\item{U}{Upper Cholesky factor of dimension `p` \eqn{\times} `p`}

\item{x}{Any `p`-long vector}
}
\value{
Upper Cholesky factor of `U'U + xx'`
}
\description{
This function calculates the rank-1 update of upper Cholesky factor `U`, i.e.
the upper Cholesky factor of matrix `U'U + xx'`. Here, `U` is a square `p`
\eqn{\times} `p` upper-triangular matrix with positive diagonal entries, and
`x` is any arbitrary vector of length `p`. This function applies Givens
rotations to triangularize the joint matrix `(U x)`. Due to cache latency,
this function will generally be slower than `cholupL`.
}
