% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lsolve.R
\name{Lsolve}
\alias{Lsolve}
\title{Lower Triangular Solver}
\usage{
Lsolve(L, x)
}
\arguments{
\item{L}{Lower-triangular matrix of dimension `p` \eqn{\times} `p`}

\item{x}{Any vector of length `p`.}
}
\description{
This function is a custom implementation of the default execution of 
`forwardsolve`, evaluating \eqn{L^{-1}x} where `L` is lower-triangular.
}
