% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LtoU.R
\name{LtoU}
\alias{LtoU}
\title{Fast L to U Transpose}
\usage{
LtoU(L)
}
\arguments{
\item{L}{Any square matrix such that its lower-triangular part (including
main diagonal) will be transposed to create an upper-triangular matrix.}
}
\value{
An upper-triangular square matrix whose non-zero entries are the
lower-triangular part of `L`.
}
\description{
This function copies the lower-triangular part of square matrix `L`
(including the main diagonal) into the upper-triangular portion of output
matrix `U`, other values zero. For `L` that is only lower-triangular, this is
effectively a faster transpose operation.
}
