% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trisolve.R
\name{trisolve}
\alias{trisolve}
\title{Triangular Solver}
\usage{
trisolve(CF, X, lower = TRUE)
}
\arguments{
\item{CF}{Triangular matrix of dimension `n` \eqn{\times} `n`}

\item{X}{Any matrix of dimension `n` \eqn{\times} `p`, or vector of length `n`}

\item{lower}{Boolean, dictating whether `CF` is lower-triangular
(`TRUE`, default) or upper-triangular (`FASLE`).}
}
\value{
Solution for `Y` in linear system `(CF)Y = X`, i.e. \eqn{(CF)^{-1}X}
}
\description{
This function is a custom implementation of the base `R` functions 
`forwardsolve` (when `lower = TRUE`) and `backsolve` (when `lower = FALSE`),
evaluating \eqn{(CF)^{-1}X}.
}
\examples{
# Generate positive definite matrix
n = 20
M = diag(n) + tcrossprod(matrix(rnorm(n * 2), n, 2))

# Calculate Cholesky factors
U = chol(M)
L = t(U)

# Generate RHS matrix for linear system
p = 5
X = matrix(rnorm(n * p), n, p)

# Calculate triangular solves
Lx = trisolve(L, X, lower = TRUE)
Ux = trisolve(U, X, lower = FALSE)

# Compare to forwardsolve and backsolve
max(abs(Lx - forwardsolve(L, X)))
max(abs(Ux - backsolve(U, X)))
}
