% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trisolveX.R
\name{trisolveX}
\alias{trisolveX}
\title{Triangular Solver with Matrix Argument}
\usage{
trisolveX(CF, X, lower = TRUE)
}
\arguments{
\item{CF}{Triangular matrix of dimension `p` \eqn{\times} `p`}

\item{lower}{Boolean, dictating whether `CF` is lower-triangular
(`TRUE`, default) or upper-triangular (`FASLE`).}

\item{x}{Any vector of length `p`.}
}
\value{
Solution for `y` in linear system `(CF)y = x`, i.e. \eqn{(CF)^{-1}x}
}
\description{
This function is a custom implementation of the default execution of 
`forwardsolve` (when `lower = TRUE`) or `backsolve` (when `lower = FALSE`),
evaluating \eqn{(CF)^{-1}x}.
}
\examples{
# Generate positive definite matrix
p = 20
M = diag(p) + tcrossprod(matrix(rnorm(p * 2), p, 2))

# Calculate Cholesky factors
U = chol(M)
L = t(U)

# Generate vector for linear system
x = rnorm(p)

# Calculate triangular solves
Lx = trisolve(L, x, lower = TRUE)
Ux = trisolve(U, x, lower = FALSE)

# Compare to forwardsolve and backsolve
max(abs(Lx - forwardsolve(L, x)))
max(abs(Ux - backsolve(U, x)))
}
