% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usolve.R
\name{Usolve}
\alias{Usolve}
\title{Upper Triangular Solver}
\usage{
Usolve(U, x)
}
\arguments{
\item{U}{Upper-triangular matrix of dimension `p` \eqn{\times} `p`}

\item{x}{Any vector of length `p`.}
}
\description{
This function is a custom implementation of the default execution of 
`backsolve`, evaluating \eqn{U^{-1}x} where `U` is upper-triangular.
}
