% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtoL.R
\name{UtoL}
\alias{UtoL}
\title{Fast U to L Transpose}
\usage{
UtoL(U)
}
\arguments{
\item{U}{Any square matrix such that its upper-triangular part (including
main diagonal) will be transposed to create a lower-triangular matrix.}
}
\value{
A lower-triangular square matrix whose non-zero entries are the
upper-triangular part of `U`.
}
\description{
This function copies the upper-triangular part of square matrix `U`
(including the main diagonal) into the lower-triangular portion of output
matrix `L`, other values zero. For `U` that is only upper-triangular, this is
effectively a faster transpose operation.
}
