% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choldrop.R
\name{choldrop}
\alias{choldrop}
\title{Cholesky Row/Column Drop}
\usage{
choldrop(CF, k, lower = TRUE)
}
\arguments{
\item{CF}{Cholesky factor of dimension `p` \eqn{\times} `p`, such that e.g.
`(CF)(CF') = M`}

\item{k}{An index between `1` and `p` inclusive, corresponding to the
row/column of `M` to be dropped}

\item{lower}{Boolean, dictating whether `CF` is the lower Cholesky factor
(`TRUE`, default) or upper Cholesky factor (`FASLE`). Also dictates whether
the returned Cholesky factor is lower- or upper-triangular.}
}
\value{
Cholesky factor of `M[-k, -k]`, i.e. of `M` with the
\eqn{k^{th}} row/column removed
}
\description{
This function calculates the Cholesky factor of matrix `M[-k, -k]`,
i.e. of positive definite `M` with the \eqn{k^{th}} row/column removed,
provided Cholesky factor `CF` (e.g. `M = (CF)(CF')`) is known. The user may
specify whether `CF` is lower- or upper-triangular; due to cache latency,
this function will generally be faster when `CF` is the lower Cholesky
factor.
}
\examples{
# Generate positive definite matrix
p = 20
M = diag(p) + tcrossprod(matrix(rnorm(p * 2), p, 2))

# Take out a row/column of M
k = 6
Mk = M[-k, -k]

# Calculate Cholesky factor of M
U = chol(M)

# Calculate Cholesky factor removing k-th row/column
Uk = choldrop(U, k, lower = FALSE)

# Check against Cholesky factor of Uk
max(abs(Uk - chol(Mk)))
}
