% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholup.R, R/cholupL.R
\name{cholupL}
\alias{cholupL}
\title{Cholesky Rank-1 Update}
\usage{
cholupL(L, x)

cholupL(L, x)
}
\arguments{
\item{L}{Lower Cholesky factor of dimension `p` \eqn{\times} `p`}

\item{x}{Any `p`-long vector}

\item{CF}{Cholesky factor of dimension `p` \eqn{\times} `p`}
}
\value{
Cholesky factor of (e.g.) `(CF)(CF') + xx'`, where `(CF)(CF')` is
the positive definite matrix associated with `CF`.

Lower Cholesky factor of `LL' + xx'`
}
\description{
This function calculates the rank-1 update of Cholesky factor `CF`, e.g.
the Cholesky factor of matrix `(CF)(CF') + xx'`. Here, `CF` is a square `p`
\eqn{\times} `p` Cholesky factor, and `x` is any arbitrary vector of length
`p`. The user may specify whether `CF` is lower-triangular (in which case
`lower = TRUE`) or upper-triangular (in which case `lower = FALSE`). This
function applies Givens rotations to triangularize the joint matrix `(CF x)`;
due to cache latency, it will generally be faster when `CF` is the lower
Cholesky factor.

This function calculates the rank-1 update of lower Cholesky factor `L`, i.e.
the lower Cholesky factor of matrix `LL' + xx'`. Here, `L` is a square `p`
\eqn{\times} `p` lower-triangular matrix with positive diagonal entries, and
`x` is any arbitrary vector of length `p`. This function applies Givens
rotations to triangularize the joint matrix `(L x)`. Due to cache latency,
this function will generally be faster than `cholupU`.
}
