% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trisolve.R
\name{trisolve}
\alias{trisolve}
\title{Triangular Solver}
\usage{
trisolve(CF, x, lower = TRUE)
}
\arguments{
\item{CF}{Triangular matrix of dimension `p` \eqn{\times} `p`}

\item{x}{Any vector of length `p`.}

\item{lower}{Boolean, dictating whether `CF` is lower-triangular
(`TRUE`, default) or upper-triangular (`FASLE`).}
}
\description{
This function is a custom implementation of the default execution of 
`forwardsolve` (when `lower = TRUE`) or `backsolve` (when `lower = FALSE`),
evaluating \eqn{(CF)^{-1}x}.
}
